requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
atlassian-python-api==3.41.8
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
coverage==7.4.0
Deprecated==1.2.14
distro==1.9.0
Django==5.0.1
django-debug-toolbar==4.2.0
exceptiongroup==1.2.0
fastembed==0.1.1
flatbuffers==23.5.26
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
jmespath==1.0.1
mpmath==1.3.0
numpy==1.26.3
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.3
openai==1.7.2
packaging==23.2
portalocker==2.8.2
protobuf==4.25.2
psycopg==3.1.17
psycopg-binary==3.1.17
pydantic==2.5.3
pydantic_core==2.14.6
qdrant-client==1.7.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
tiktoken==0.5.2
tokenizers==0.13.3
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
wrapt==1.16.0
python-dotenv==0.19.1
